Crypto Currency Exchange /Subscriber/requirements.txt
Flask==1.1.2
pandas==1.3.3
Flask-Jsonpify==1.5.0
kafka-python==2.0.2